fix: bug where restoring value when leaving minimalist mode
When leaving minimalist mode we restore all values. Sometimes a value is "1"
or "0" (to represent booleans) but cannot be restored with "set $OPTION=1" we
must then use "set $OPTION" or "set no$OPTION".
E.g: ruler is enabled (set ruler). When we insert minimalist mode ruler
original value is saved with the value "1". When we restore original options
we use "set ruler" instead of "set ruler=1" (which will crash).